refactor: removing prisma client from app.ts, putting it into lib folder and changing imports
